+++ EssentialFeed/Feed Presentation/FeedErrorViewModel.swift
@@ -1,0 +1,20 @@
+//
+//  FeedErrorViewModel.swift
+//  EssentialFeed
+//
+//  Created by Julius on 01/08/2022.
+//
+
+import Foundation
+
+public struct FeedErrorViewModel {
+    public let message: String?
+    
+    static var noError: FeedErrorViewModel {
+        return FeedErrorViewModel(message: nil)
+    }
+    
+    static func error(message: String) -> FeedErrorViewModel {
+        return FeedErrorViewModel(message: message)
+    }
+}
